requirements.txt
requests>=2.31.0
pymysql>=1.1.0
psycopg2-binary>=2.9.0
oracledb>=2.0.0
openpyxl>=3.1.0
